Move some more stuff to better places
